Use the storage backend to setup/teardown data.
If we directly attack the filesystem, we'll only ever test against a
filesystem storage backend (and not even on all of them).

Also use a smaller test image, because an 88K one takes an age to
upload for every single test when working with a remote backend.

Finally, when testing URLs, use the storage backend to generate them
rather than concatenating MEDIA_URL onto anything (since that will
only work on storage backends that don't modify paths).